+++ lib/util/operationReducer.ts
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-import * as it from "@ts-common/iterator"
 import * as sm from "@ts-common/string-map"
 
 import { ModelValidationError } from "./modelValidationError"
@@ -18,9 +17,12 @@ export function operationReducer({
   scenarios
 }: OperationResultScenarios): Iterable<ModelValidationError> {
   const scenariosEntries = sm.entries(scenarios)
-  const invalidScenarios = it.filter(scenariosEntries, ([_, scenario]) => !scenario.isValid)
-  const result = it.flatMap(invalidScenarios, ([scenarioName, scenario]) =>
-    scenarioReducer(scenarioName, scenario, operationId)
-  )
-  return result
+  const invalidScenarios = scenariosEntries.filter(entry => {
+    const [, scenario] = entry
+    return !scenario.isValid
+  })
+  return invalidScenarios.flatMap(entry => {
+    const [scenarioName, scenario] = entry
+    return scenarioReducer(scenarioName, scenario, operationId)
+  })
 }